Adding reward tables
Reward tables for Minecolonies quests. The quantities on the tables themselves will likely have to be tweaked, and some items will need to be swapped in or out. It may be useful to weight some of the rewards, since they are used both as random rewards and choice rewards.

Intro reward : Choice of 3 parrots, 1 wolf, or 1 cat (all untamed). Mostly a placeholder, on my personal server there was also a joke option (and a hidden joke quest to play a server-wide announcement when someone selected that option). The joke option and associated quest have been removed. A more useful intro reward would probably be a small quantity of leather, so players can make the resource scroll and clipboard.

Reward II: Building block rewards. Minecolonies is somewhat infamous for its high consumption of basic resources in the form of building blocks. I have repeatedly observed players who are interested in Minecolonies drop the mod because this serves as such a high barrier to entry - the colony requires a builder's hut, food source, restaurant, and housing (usually a tavern) before it can even begin to be self-sufficient. I have used building blocks (a half stack of smooth stone or clay, two stacks of sand or gravel, or one stack of cobblestone or a vanilla log) as the primary reward throughout this quest chain; a random reward for starting a building and a choice reward for completing a building. My hope is that players are able to short circuit the grind of gathering resources (particularly if they set up bonsai farms for their colony's main wood) in favor of crafting each of the many blocks the schematics actually need. The mod becomes a lot more accessible to new players if early building tasks have minimal resource grinding and their primary activity is turning those resources into fences, stairs, etc. This also makes the step up to manufacturing professions more rewarding - once the player has a carpenter who can produce stairs and fences, they're free to focus on more enjoyable aspects of the mod over micromanaging the production and fulfillment of every request involved in building.

Reward III: Item rewards. These are smaller in quantity, but again have been chosen based on reducing the amount of grinding necessary to get the colony on its feet. Given in small enough quantities that they should be "just enough" for the schematics that require them, meaning that they should end up being used up pretty much as soon as they're received instead of sitting around in a chest somewhere. Selected on the basis that if a player has to spend 20-40 minutes early on roaming around looking for cows for leather, or trekking to another biome for a few sugar cane to start a farm, they'll come back to a colony full of citizens who work extremely slowly because they've been homeless and hungry for several weeks. (Or worse, a single skeleton wipes out the entire colony population in the player's absence)

Reward IV: Base seeds/saplings. Vanilla saplings, or vanilla crop seeds. Intended for players who need to start a tree farm to obtain a source of whichever wood is second-most common in their selected schematic style. This can probably be trimmed in quantity, given how fast it is to grow a large tree farm from bonsai pots. Perhaps including a single bonsai pot along with 2-3 saplings of the selected variety would be more balanced, with the quest text encouraging the player to read the tip about their use.

Reward V: Animal rewards. By the time a player decides to construct a ranch building, they probably have observed the related animal nearby. This is primarily intended to allow a colony to start a "non native" ranch of some variety. Horse, goat, and llama eggs are included despite no villager hut making use of them purely for player use. As the individual quests only track the creation of the hut items and not the completion (completion of a building being tracked by advancements currently - I know that the minecolonies API does have a building request complete event which contains the building type and upgrade level, but did not look into how to track this in FTB quests), the animal hut quests do not reward animals. This is so that a player cannot construct all of the hut types in one fell swoop for a reward without ever finishing them. There is an advancement (and quest) for constructing all 5 animal huts, but the delayed gratification of completing five construction projects before getting the corresponding animals may be too long a wait for too little benefit. It might be best to provide two eggs of only the corresponding hut type for each quest (preferably on hut completion), and then a different, larger reward for completing all 5.

Rewards VI: A small choice of food seeds or food tree sapling. This may be a less useful reward, but I think the saplings are an interesting alternative to a traditional farm.

Rewards VII: Minecolonies scrolls. These are probably considerably more useful as a reward before the colony is able to produce them through the enchanter. The most useful one early game is the highlight scroll, which speeds up the builder when used on the builder's hut.