Minecolonies quest chapter
Some background about this chapter:
The design goal of this quest chapter was to tutorialize the basic progression of Minecolonies by providing useful advice or information from the wiki and my own experiences with the mod, while providing rewards which reduce the resource grind the mod is known for. The process of starting a colony is by itself very easy and straightforward - you construct a supply ship or supply camp, place it, then place your town hall schematic and your first set of colonists spawn in. However, starting a *functional* colony, and getting it to the point of self-sufficiency, is a *much* more involved process and I believe is ultimately the reason many players drop the mod before they start seeing a return on their time and resource investment. Starting a colony and spawning colonists is extremely easy, but the first step, constructing a builder's hut, requires a large amount of resources and understanding the way the request system works. Depending on the supply camp style selected and the builder's hut style selected, players can often deconstruct the majority of the camp to obtain most of the blocks the builder needs. A first time player may not know this, and even if they do, might take an hour or more to babysit the builder through the proces. By the time the builder's hut is complete, the colonists will likely be complaining about food, shelter, and safety. If the player decides to gather the resources themselves for their next few construction projects, typically housing (usually a tavern), a food source, and a gatherer such as a miner or forester, they're likely to return to dead or unhappy colonists, hours of micromanaging the builder, and low level gatherers who need a constant supply of tools and other supplies in order to break only a few blocks per minute. This problem is compounded when the player tries to construct a mine or forester first to help with the cost of construction - another hour of babysitting the builder only to have the new worker produce far less material than is needed for the next construction project. When players consider the amount of time and resources they've invested to get to this point compared to the lack of benefit they personally gain from the colony, they're likely to give up and focus on another mod with a more immediate return like blood magic or the industrial series.

With these experiences in mind, the quests are written and ordered to guide players through the mod in a similar manner to the Minecolonies "Getting Started" guide, with some other tips they would otherwise need to look up on the wiki or learn through experience. Getting housing, food, and basic gathering are prioritized over diving directly into item production, military, or research. The rewards provided are intended to minimize but not eliminate the early grind for wood and stone, as well as providing "just enough" of other items like leather and wool. Playtesting by myself and on a server has indicated that these are useful early on (though the quantities require tweaking and the item-type rewards could use another pass), but these rewards become progressively less useful as the colony develops. While it would require additional research (into structurize blueprints and block counts) and development (reward tables and itemsets), I think it would ultimately be much more rewarding to offer sets of blocks as rewards with players selecting from the structurize theme types - fortress giving dark oak and stone brick, dwarven giving more smooth stone, etc.

Quests are typically divided into two parts - the construction of a hut block/building item, which offers a token item or random building block reward, and the completion of the building, which usually offers a player choice of the same building blocks or sometimes other items. Building completion is tracked by advancements - not all buildings have an associated advancement, but it should be possible to either track the same triggers or create new advancements. The trigger used for the building completion advancements in minecolonies is minecolonies:complete_build_request, which contains conditions for hut_name (name of the building) and level (upgrade level of the building). A new building will have a level of 1, but upgrading a building to a higher level also fires the same complete_build_request event, with the new level passed in. This is how the advancements (and quests) for upgrading the builder's hut, town hall, etc. work.

In terms of what would be useful for additional quests, building completion quests for those which don't have them would be useful. At the moment, animal herder huts only have a "make the item" quest. I opted not to reward the corresponding animals for these quests, to discourage players from simply crafting all the building items and never building them in order to reap the rewards. A breeding pair of the associated animal for each hut would be a fine reward on building completion.

As far as gating, none of the mod content is actually gated beyond buildings which require university research or another building to be upgraded to a certain point. The quests are invisible until discovery purely to discourage players from doing something like building a guard tower before housing, or neglecting to upgrade their buildings. The quest to upgrade the builder's hut becomes visible as soon as they get their first additional colonist, either a child or hired visitor from the tavern, and the quest text explains that they need to upgrade the builder's hut in order to upgrade other buildings. The guard tower quest is behind the 5 population quest, and the university quest is behind the 10 population quest. Players can obviously build these things before the quest is visible, though this does leave the problem of having to craft another "building item" to progress beyond any quests unlocked after the buildings in question are already constructed. While the advancement parts should complete without issue if the advancement is done before the quest becomes visible, the item-linked quests will not.

Currently, military achivement and population count quests are devoid of text. Their rewards were also chosen without much care.
It would be nice to see either one of the later military or population quests reward some contribution towards the creation of the ATM star or some creative item.